alert_processor/src/requirements.txt
firebase-admin==7.0.0
google-cloud-pubsub==2.31.0
google-cloud-firestore
